+++ exe/Main.hs
@@ -5,6 +5,7 @@
 {-# LANGUAGE OverloadedRecordDot #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE QuasiQuotes #-}
+{-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE TypeApplications #-}
 {-# OPTIONS_GHC -Wno-missing-signatures #-}
 
@@ -18,19 +19,23 @@ import Slop hiding (clear, draw)
 import Slop.GPU hiding (drawMesh, output, pass, postProcess, render, withShader)
 import Slop.Pipeline.Graph
 import Slop.SDL.Raw (GPUDevice(..), GPUTexture(..))
-import Spirdo.Wesl
+import Spirdo.Wesl.Reflection
   ( ToUniform
-  , prepareShader
+  , defaultCompileOptions
+  , imports
   , shaderSpirv
+  , spirv
   , wesl
   )
-import qualified Spirdo.Wesl as W
+import qualified Spirdo.Wesl.Reflection as W
 import Spirdo.Wesl.Inputs
-  ( ShaderInputs(..)
+  ( InputsError(..)
+  , ShaderInputs
   , StorageTextureInput(..)
   , StorageTextureHandle(..)
   , UniformInput(..)
-  , inputsFromPrepared
+  , inputsFor
+  , inputsStorageTextures
   , orderedUniforms
   , storageTexture
   , uniform
@@ -57,7 +62,7 @@ data SliceParams = SliceParams
 instance ToUniform SliceParams
 
 demoShader =
-  [wesl|
+  $(spirv defaultCompileOptions imports [wesl|
 // spirdo:sampler=combined
 struct SlopGlobals {
   time: f32,
@@ -84,10 +89,10 @@ fn main(@location(0) uv: vec2<f32>, @location(1) inColor: vec4<f32>) -> @locatio
 
   return vec4(color.r * inColor.r, color.g * inColor.g, color.b * inColor.b, inColor.a);
 }
-|]
+|])
 
 comboShader =
-  [wesl|
+  $(spirv defaultCompileOptions imports [wesl|
 // spirdo:sampler=combined
 @group(2) @binding(0) var spriteTex: texture_2d<f32>;
 @group(2) @binding(2) var spriteSamp: sampler;
@@ -102,10 +107,10 @@ fn main(@location(0) uv: vec2<f32>, @location(1) inColor: vec4<f32>) -> @locatio
   let mixed = base.rgb * (boost + noise.rgb * boost);
   return vec4(mixed.r, mixed.g, mixed.b, base.a) * inColor;
 }
-|]
+|])
 
 slice3dShader =
-  [wesl|
+  $(spirv defaultCompileOptions imports [wesl|
 // spirdo:sampler=combined
 struct SliceParams {
   time: f32;
@@ -138,10 +143,10 @@ fn main(@builtin(global_invocation_id) gid: vec3<u32>) {
   let col = mix(a, b, vec4(t, t, t, t));
   textureStore(out_tex, vec2(i32(gid.x), i32(gid.y)), vec4(col.x, col.y, col.z, 1.0));
 }
-|]
+|])
 
 computeShader =
-  [wesl|
+  $(spirv defaultCompileOptions imports [wesl|
 // spirdo:sampler=combined
 struct Params {
   time: f32;
@@ -166,7 +171,7 @@ fn main(@builtin(global_invocation_id) gid: vec3<u32>) {
   let col = vec3(0.15 + wave * 0.6, 0.2 + wave * 0.4, 0.85 - wave * 0.5);
   textureStore(out_tex, vec2(i32(gid.x), i32(gid.y)), vec4(col.x, col.y, col.z, 1.0));
 }
-|]
+|])
 
 main :: IO ()
 main = do
@@ -181,17 +186,12 @@ main = do
 
   runWindow cfg $ do
     let load spec = loadAssetAsync spec >>= awaitAsset
-    let prepare label shader = either (throwSpirdoError "prepare shader" label) pure (prepareShader shader)
-    let inputs label prepared builder = either (throwSpirdoError "build shader inputs" label) pure (inputsFromPrepared prepared builder)
+    let inputs label shader builder = either (throwInputsError label) pure (inputsFor shader builder)
 
-    computePrepared <- prepare "computeShader" computeShader
-    slicePrepared <- prepare "slice3dShader" slice3dShader
-    demoPrepared <- prepare "demoShader" demoShader
-    comboPrepared <- prepare "comboShader" comboShader
-    computeLayout <- either throwSlop pure (reflectionFromPrepared "computeShader" computePrepared)
-    sliceLayout <- either throwSlop pure (reflectionFromPrepared "slice3dShader" slicePrepared)
-    demoLayout <- either throwSlop pure (reflectionFromPrepared "demoShader" demoPrepared)
-    comboLayout <- either throwSlop pure (reflectionFromPrepared "comboShader" comboPrepared)
+    computeLayout <- either throwSlop pure (reflectionFromShader "computeShader" computeShader)
+    sliceLayout <- either throwSlop pure (reflectionFromShader "slice3dShader" slice3dShader)
+    demoLayout <- either throwSlop pure (reflectionFromShader "demoShader" demoShader)
+    comboLayout <- either throwSlop pure (reflectionFromShader "comboShader" comboShader)
     computeDesc <- either throwSlop pure (computeDescFromReflection (shaderSpirv computeShader) (threads3D 8 8 1) computeLayout)
     sliceDesc <- either throwSlop pure (computeDescFromReflection (shaderSpirv slice3dShader) (threads3D 8 8 1) sliceLayout)
 
@@ -270,13 +270,13 @@ main = do
       let sliceParams = SliceParams frame.time (fromIntegral volumeSize) (fromIntegral volumeSize) (fromIntegral volumeSize)
 
       computeInputs <-
-        inputs "compute inputs" computePrepared
+        inputs "computeShader" computeShader
           ( storageTexture @"out_tex" (toStorageHandle computeTex)
               <> uniform @"params" computeParams
           )
 
       sliceInputs <-
-        inputs "slice inputs" slicePrepared
+        inputs "slice3dShader" slice3dShader
           ( storageTexture @"noiseVolume" (toStorageHandle noiseVolume)
               <> storageTexture @"out_tex" (toStorageHandle noiseSliceTex)
               <> uniform @"params" sliceParams
@@ -345,14 +345,20 @@ storageTextureFromHandle (StorageTextureHandle value) =
     , textureDepth = 1
     }
 
-throwSpirdoError :: T.Text -> T.Text -> String -> WindowM a
-throwSpirdoError operation shaderName detail =
-  throwSlop (SlopShaderFailure operation shaderName Nothing (T.pack detail))
+throwInputsError :: T.Text -> InputsError -> WindowM a
+throwInputsError shaderName inputError =
+  throwSlop
+    ( SlopShaderFailure
+        "build shader inputs"
+        shaderName
+        (T.pack <$> inputError.ieBinding)
+        (T.pack inputError.ieMessage)
+    )
 
-reflectionFromPrepared :: T.Text -> W.PreparedShader iface -> SlopResult ReflectedShaderLayout
-reflectionFromPrepared shaderName prepared = do
-  bindings <- mapM (reflectedBinding shaderName) prepared.psPlan.bpBindings
-  shaderLayoutFromReflection shaderName (reflectedStage prepared.psStage) bindings
+reflectionFromShader :: T.Text -> W.Shader mode iface -> SlopResult ReflectedShaderLayout
+reflectionFromShader shaderName shader = do
+  bindings <- mapM (reflectedBinding shaderName) (W.shaderPlan shader).bpBindings
+  shaderLayoutFromReflection shaderName (reflectedStage (W.shaderStageCached shader)) bindings
 
 reflectedStage :: W.ShaderStage -> ShaderStage
 reflectedStage stage =
@@ -372,18 +378,35 @@ reflectedBinding shaderName binding = do
     }
 
 reflectedBindingType :: T.Text -> W.BindingInfo -> SlopResult ReflectedBindingType
-reflectedBindingType shaderName binding
-  | W.isUniformKind binding.biKind =
+reflectedBindingType shaderName binding =
+  case binding.biKind of
+    W.BUniform ->
       case uniformLayoutSize binding.biType of
         Just byteSize -> Right (ReflectedUniform byteSize)
         Nothing -> Left (unsupported "uniform reflection did not contain a host-shareable byte size")
-  | W.isSamplerKind binding.biKind = Right ReflectedSampler
-  | W.isTextureKind binding.biKind = Right ReflectedSampledTexture
-  | W.isStorageBufferKind binding.biKind = Right ReflectedStorageBuffer
-  | W.isStorageTextureKind binding.biKind =
-      ReflectedStorageTexture <$> reflectedStorageAccess binding.biType
-  | otherwise = Left (unsupported "binding kind is not supported")
+    W.BSampler -> Right ReflectedSampler
+    W.BSamplerComparison -> Right ReflectedSampler
+    W.BTexture1D -> Right ReflectedSampledTexture
+    W.BTexture1DArray -> Right ReflectedSampledTexture
+    W.BTexture2D -> Right ReflectedSampledTexture
+    W.BTexture2DArray -> Right ReflectedSampledTexture
+    W.BTexture3D -> Right ReflectedSampledTexture
+    W.BTextureCube -> Right ReflectedSampledTexture
+    W.BTextureCubeArray -> Right ReflectedSampledTexture
+    W.BTextureMultisampled2D -> Right ReflectedSampledTexture
+    W.BTextureDepth2D -> Right ReflectedSampledTexture
+    W.BTextureDepth2DArray -> Right ReflectedSampledTexture
+    W.BTextureDepthCube -> Right ReflectedSampledTexture
+    W.BTextureDepthCubeArray -> Right ReflectedSampledTexture
+    W.BTextureDepthMultisampled2D -> Right ReflectedSampledTexture
+    W.BStorageRead -> Right ReflectedStorageBuffer
+    W.BStorageReadWrite -> Right ReflectedStorageBuffer
+    W.BStorageTexture1D -> storageTextureType
+    W.BStorageTexture2D -> storageTextureType
+    W.BStorageTexture2DArray -> storageTextureType
+    W.BStorageTexture3D -> storageTextureType
   where
+    storageTextureType = ReflectedStorageTexture <$> reflectedStorageAccess binding.biType
     unsupported detail =
       SlopShaderFailure "reflect shader" shaderName (Just (T.pack binding.biName)) detail
 
@@ -422,5 +445,5 @@ computeBindingsFromInputs layout inputs =
       [ ComputeStorageTextureRWNamed
           (T.pack entry.storageTextureName)
           (storageTextureFromHandle entry.storageTextureHandle)
-      | entry <- inputs.siStorageTextures
+      | entry <- inputsStorageTextures inputs
       ]